requirements.txt
amqp==2.5.0
billiard==3.6.0.0
celery==4.3.0
Django==2.2.3
django-extensions==2.1.9
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
kombu==4.6.3
PyJWT==1.7.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
vine==1.3.0
